backend/requirements.txt
fastapi
uvicorn[standard]
pytesseract
opencv-python
pillow
httpx
aiofiles
huggingface-hub
python-dotenv
